Added new service to handler all tokens
